refactor: вынес исключения в одно место
