+++ hw-2/src/main/java/ru/dankoy/core/csvreader/CsvReaderImpl.java
@@ -8,6 +8,7 @@ import java.util.List;
 import java.util.Objects;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
+import ru.dankoy.core.exceptions.CsvReaderException;
 
 @Component
 public class CsvReaderImpl implements CsvReader {
