+++ hw-2/src/main/java/ru/dankoy/core/dao/csv/CsvQuestionsDaoImpl.java
@@ -7,6 +7,7 @@ import ru.dankoy.core.csvreader.CsvReader;
 import ru.dankoy.core.dao.QuestionsDao;
 import ru.dankoy.core.domain.Answer;
 import ru.dankoy.core.domain.Question;
+import ru.dankoy.core.exceptions.QuestionsDaoException;
 
 /**
  * @author Dankoy
@@ -29,34 +30,39 @@ public class CsvQuestionsDaoImpl implements QuestionsDao {
     List<String[]> rows = csvReader.read();
 
     List<Question> questions = new ArrayList<>();
-    // get amount of questions
-    for (String[] row : rows) {
 
-      var questionId = Long.parseLong(row[0]);
-      var questionText = row[1];
+    try {
+      // get amount of questions
+      for (String[] row : rows) {
 
-      // parse answers. may have different amount
-      var correctAnswerId = -1;
-      List<Answer> answers = new ArrayList<>();
-      var answerId = 1;
-      for (int i = 2; i < row.length; i += 2) {
+        var questionId = Long.parseLong(row[0]);
+        var questionText = row[1];
 
-        if (!row[i].isEmpty() || !row[i].isBlank()) {
-          answers.add(new Answer(answerId, row[i]));
+        // parse answers. may have different amount
+        var correctAnswerId = -1;
+        List<Answer> answers = new ArrayList<>();
+        var answerId = 1;
+        for (int i = 2; i < row.length; i += 2) {
 
-          var isCorrect = Boolean.parseBoolean(row[i + 1]);
+          if (!row[i].isEmpty() || !row[i].isBlank()) {
+            answers.add(new Answer(answerId, row[i]));
 
-          if (isCorrect) {
-            correctAnswerId = answerId;
+            var isCorrect = Boolean.parseBoolean(row[i + 1]);
+
+            if (isCorrect) {
+              correctAnswerId = answerId;
+            }
           }
+          answerId++;
+
         }
-        answerId++;
+
+        var question = new Question(questionId, questionText, answers, correctAnswerId);
+        questions.add(question);
 
       }
-
-      var question = new Question(questionId, questionText, answers, correctAnswerId);
-      questions.add(question);
-
+    } catch (Exception e) {
+      throw new QuestionsDaoException(e);
     }
 
     return questions;
